refactor: 에러 및 로딩 상태를 관리하는 커스텀 훅 분리
